# src/swagger-new.json
{
  "swagger": "2.0",
  "info": {
    "title": "PCTE",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3000/api",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/tar/": {
      "get": {
        "tags": [
          "Tar"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "string",
                  "example": "0.0000"
                },
                "valorTarifario": {
                  "type": "string",
                  "example": "0.0000"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Tar"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "string",
                  "example": "0.0000"
                },
                "valorTarifario": {
                  "type": "string",
                  "example": "0.0000"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tar/{id}": {
      "get": {
        "tags": [
          "Tar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "string",
                  "example": "0.0000"
                },
                "valorTarifario": {
                  "type": "string",
                  "example": "0.0000"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Tar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "idPotencia": {
                  "type": "number",
                  "example": 1
                },
                "idTarifario": {
                  "type": "number",
                  "example": 1
                },
                "valorPotencia": {
                  "type": "string",
                  "example": "0.0000"
                },
                "valorTarifario": {
                  "type": "string",
                  "example": "0.0000"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Tar"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "affected": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tarifario/": {
      "get": {
        "tags": [
          "Tarifario"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "tipo": {
                  "type": "string",
                  "example": "Simples"
                },
                "valor": {
                  "type": "string",
                  "example": "Fixo"
                },
                "contrato": {
                  "type": "string",
                  "example": "Particular"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Tarifario"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "tipo": {
                  "type": "string",
                  "example": "Simples"
                },
                "valor": {
                  "type": "string",
                  "example": "Fixo"
                },
                "contrato": {
                  "type": "string",
                  "example": "Particular"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tarifario/{id}": {
      "get": {
        "tags": [
          "Tarifario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "tipo": {
                  "type": "string",
                  "example": "Simples"
                },
                "valor": {
                  "type": "string",
                  "example": "Fixo"
                },
                "contrato": {
                  "type": "string",
                  "example": "Particular"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Tarifario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "tipo": {
                  "type": "string",
                  "example": "Simples"
                },
                "valor": {
                  "type": "string",
                  "example": "Fixo"
                },
                "contrato": {
                  "type": "string",
                  "example": "Particular"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Tarifario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "raw": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "affected": {
                  "type": "number",
                  "example": 1
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}